Assignment number 1 problem statement 5 solution
In this problem statement I have shown serialization and deserialization. As stated in the assignment. I have created a Student class and an Address class in it.  Along with this, screenshots have also been put as mentioned in it.  And a file named Assignment Description is created.  In which complete details of the assignment are included.